+++ src/gui/update_dialog.py
@@ -56,7 +56,7 @@ class UpdateAvailableDialog(QDialog):
         buttons = QDialogButtonBox(QDialogButtonBox.Yes | QDialogButtonBox.No)
         yes_button = buttons.button(QDialogButtonBox.Yes)
         if yes_button is not None:
-            yes_button.setText('Download')
+            yes_button.setText('Download and Install')
         no_button = buttons.button(QDialogButtonBox.No)
         if no_button is not None:
             no_button.setText('Later')